+++ source/Internal/Framework/Module/Install/Service/ModuleConfigurationInstaller.php
@@ -18,6 +18,9 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\MetaData\Dao\ModuleConfi
 use OxidEsales\EshopCommunity\Internal\Framework\Module\MetaData\Exception\InvalidMetaDataException;
 use OxidEsales\EshopCommunity\Internal\Transition\Utility\BasicContextInterface;
 use Webmozart\PathUtil\Path;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\Service\{
+    ModuleConfigurationMergingServiceInterface
+};
 
 class ModuleConfigurationInstaller implements ModuleConfigurationInstallerInterface
 {
@@ -87,6 +90,22 @@ class ModuleConfigurationInstaller implements ModuleConfigurationInstallerInterf
         foreach ($projectConfiguration->getShopConfigurations() as $shopConfiguration) {
             if ($shopConfiguration->hasModuleConfiguration($moduleConfiguration->getId())) {
                 $shopConfiguration->deleteModuleConfiguration($moduleConfiguration->getId());
+            }
+        }
+
+        $this->projectConfigurationDao->save($projectConfiguration);
+    }
+
+    /**
+     * @param string $moduleId
+     */
+    public function uninstallById(string $moduleId): void
+    {
+        $projectConfiguration = $this->projectConfigurationDao->getConfiguration();
+
+        foreach ($projectConfiguration->getShopConfigurations() as $shopConfiguration) {
+            if ($shopConfiguration->getModuleConfiguration($moduleId)) {
+                $shopConfiguration->deleteModuleConfiguration($moduleId);
             }
         }