+++ tests/Codeception/_data/modules/WithNamespaceAndMetadataV2/Application/Model/TestModuleNineModel.php
@@ -1,0 +1,24 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+namespace OxidEsales\EshopCommunity\Tests\Acceptance\Admin\testData\modules\Vendor1\WithNamespaceAndMetadataV2\Application\Model;
+
+/**
+ * Class TestModuleNineModel
+ *
+ * @package OxidEsales\EshopCommunity\Tests\Acceptance\Admin\testData\modules\Vendor1\WithNamespaceAndMetadataV2\Application\Model
+ */
+class TestModuleNineModel
+{
+    /**
+     * @return string
+     */
+    public function getInfo()
+    {
+        return 'TestModuleNineModel info';
+    }
+}