+++ tests/Codeception/_data/modules/test-module-1/Controller/ContentController.php
@@ -1,0 +1,11 @@
+<?php
+
+/**
+ * Copyright © OXID eSales AG. All rights reserved.
+ * See LICENSE file for license details.
+ */
+
+class ContentController extends \OxidEsales\Eshop\Application\Controller\ContentController
+{
+    
+}